third_party/WebKit/LayoutTests/inspector/sources/debugger-breakpoints/set-breakpoint-expected.txt

Tests setting breakpoints.


Running: testSetBreakpoint
Script source was shown.
Set timer for test function.
Script execution paused.
Call stack:
    0) testFunction (set-breakpoint.html:17)
Breakpoint sidebar pane 
set-breakpoint.html:17    var x = Math.sqrt(10);
Breakpoint sidebar pane 
No Breakpoints
Script execution resumed.

Running: testSetBreakpointOnTheLastLine
Call stack:
    0) oneLineTestFunction (set-breakpoint.html:7)

Running: testSetBreakpointOnTheLastLine2
Call stack:
    0) oneLineTestFunction2 (set-breakpoint.html:11)

Running: testSetBreakpointOnTheSameLine
